Changed set_FL/BR/Z_CORR behavior + reformating
Setting a zone's corner with an invalid position will now be allowed,
but it will change the opposite corner in order to resolve the conflict.
Both corners can now be on the same spot

Changed goto_x/y/z to use a new function named clamp.
Reodered imports and removed the unused ones.